# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.DeviceToken:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_used:
        type: string
      platform:
        description: ios, android
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.Location:
    properties:
      city:
        type: string
      country:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      state:
        type: string
    type: object
...
  models.Match:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      user1:
        $ref: '#/definitions/models.User'
      user1_id:
        type: string
      user2:
        $ref: '#/definitions/models.User'
      user2_id:
        type: string
    type: object
...
  models.Photo:
    properties:
      created_at:
        type: string
      id:
        type: string
      position:
        description: Order of photos (1 = primary)
        type: integer
      url:
        type: string
      user_id:
        type: string
    type: object
...
  models.RegisterDeviceRequest:
    properties:
      platform:
        enum:
        - ios
        - android
        - web
        type: string
      token:
        type: string
    required:
    - platform
    - token
    type: object
...
  models.User:
    properties:
      age:
        type: integer
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      last_seen:
        type: string
      location:
        $ref: '#/definitions/models.Location'
      photos:
        items:
          $ref: '#/definitions/models.Photo'
        type: array
      preferences:
        $ref: '#/definitions/models.UserPrefs'
      updated_at:
        type: string
    type: object
  models.UserPrefs:
    properties:
      age_max:
        type: integer
      age_min:
        type: integer
      created_at:
        type: string
      hide_age:
        type: boolean
      hide_distance:
        type: boolean
      id:
        type: string
      interested_in:
        description: genders
        items:
          type: string
        type: array
      max_distance:
        description: in kilometers
        type: integer
      only_verified:
        type: boolean
      show_me:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@matching-api.example.com
    name: API Support
    url: https://matching-api.example.com/support
  description: A complete Go-based REST API for a dating application with real-time
    chat, sophisticated matching algorithms, and comprehensive user management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://matching-api.example.com/terms
  title: Matching API - Matching App Backend
  version: "1.0"
paths:
...
  /matches:
    get:
      consumes:
      - application/json
      description: Retrieve user's current matches with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of matches per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matches retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    matches:
                      items:
                        $ref: '#/definitions/models.Match'
                      type: array
                    page:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad request - invalid pagination
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user matches
      tags:
      - Matching
  /matches/{matchID}:
    delete:
      consumes:
      - application/json
      description: Remove/unmatch with another user
      parameters:
      - description: Match ID
        in: path
        name: matchID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Match removed successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad request - invalid match ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - not part of this match
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a match
      tags:
      - Matching
...
  /notifications/devices:
    post:
      consumes:
      - application/json
      description: Register a device token to receive push notifications
      parameters:
      - description: Device registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Device registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.DeviceToken'
              type: object
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register device for push notifications
      tags:
      - Notifications
  /notifications/devices/{tokenID}:
    delete:
      consumes:
      - application/json
      description: Remove a device token from receiving push notifications
      parameters:
      - description: Device token ID
        in: path
        name: tokenID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Device unregistered from notifications
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad request - missing token ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Device token not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unregister device from push notifications
      tags:
      - Notifications
...
  /notifications/test:
    post:
      consumes:
      - application/json
      description: Send a test notification to verify notification setup
      produces:
      - application/json
      responses:
        "200":
          description: Test notification sent
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                    notification_id:
                      type: string
                  type: object
              type: object
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send test notification
      tags:
      - Notifications
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"